requirements.txt
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
joblib>=1.3.0,<2.0.0
xgboost>=2.0.0,<3.0.0
lightgbm>=4.0.0,<5.0.0
streamlit>=1.30.0,<2.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
rapidfuzz>=3.5.0,<4.0.0
urllib3>=2.0.0,<3.0.0
